requirements.txt
# Core ADK dependencies
google-adk
python-dotenv

# LLM Integration
litellm

# Type hints (if needed for older Python versions)
typing-extensions
